refactor: moved neat config ut to top level as it is used in workflows and app
