+++ cognite/neat/app/api/configuration.py
@@ -7,8 +7,7 @@ from cognite.client import CogniteClient
 from fastapi import FastAPI
 
 from cognite import neat
-from cognite.neat.app.api.data_classes.configuration import Config, configure_logging
-from cognite.neat.config import copy_examples_to_directory, create_data_dir_structure
+from cognite.neat.config import Config, configure_logging, copy_examples_to_directory, create_data_dir_structure
 from cognite.neat.constants import PACKAGE_DIRECTORY
 from cognite.neat.utils.cdf import ServiceCogniteClient
 from cognite.neat.utils.utils import get_cognite_client_from_config, get_cognite_client_from_token
