+++ cognite/neat/app/api/routers/configuration.py
@@ -4,7 +4,8 @@ from pathlib import Path
 
 from fastapi import APIRouter
 
-from cognite.neat.app.api.configuration import NEAT_APP, Config
+from cognite.neat.app.api.configuration import NEAT_APP
+from cognite.neat.config import Config
 
 router = APIRouter()
